[refactor] hw_L51: replaced pages 'homrpage', 'sign up' into directory 'pages'
