+++ lesson_46/HW_L46/demy-app/src/App.js
@@ -1,8 +1,8 @@
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import {Header} from "./components/Header";
 import {Main} from "./components/Main";
-import {Homepage} from "./components/Homepage";
-import {SignUp} from "./components/SignUp";
+import {Homepage} from "./pages/Homepage";
+import {SignUp} from "./pages/SignUp";
 import {LINKS} from "./constants";
 
 function App() {
@@ -12,7 +12,9 @@ function App() {
         <Header />
         <Routes>
           <Route path={LINKS.homepage} element={<Main />}>
-            <Route path={LINKS.homepage} element={<Homepage />}></Route>
+            <Route path={LINKS.homepage} element={<Homepage />} />
+            <Route path={LINKS.courses} element={<Homepage />} />
+            <Route path={LINKS.profile} element={<Homepage />} />
             <Route path={LINKS.registration} element={<SignUp />} />
           </Route>
         </Routes>
