+++ lesson_46/HW_L46/demy-app/src/pages/SignUp/SignUp.jsx
@@ -1,6 +1,6 @@
 import {useNavigate, Link} from "react-router-dom";
-import {Button} from "../Button";
-import {Input} from "../Input";
+import {Button} from "../../components/Button";
+import {Input} from "../../components/Input";
 import {useForm} from "./useForm";
 import {LINKS} from "../../constants";
 import styles from "./_sign-up.module.scss";
